device from amde

+++ src/main/java/vislab/no/ntnu/vislabcontroller/webcontroller/DeviceController.java
@@ -4,12 +4,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.net.InetAddress;
+import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
@@ -45,6 +48,10 @@ public class DeviceController {
         return new ResponseEntity<>(deviceRepository.findAll(), HttpStatus.OK);
     }
 
+    @RequestMapping("/types")
+    public ResponseEntity<List<DeviceType>> getTypes(){
+        return new ResponseEntity<>(deviceTypeRepository.findAll(), HttpStatus.OK);
+    }
     @RequestMapping(value = "/getone"
             , method = RequestMethod.GET)
     public ResponseEntity<Device> getOne(@RequestParam("id") Optional<Integer> id
@@ -60,6 +67,7 @@ public class DeviceController {
         return new ResponseEntity<>(d, HttpStatus.OK);
     }
 
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @RequestMapping(value = "/add"
             , method = RequestMethod.POST
             , consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
@@ -67,22 +75,47 @@ public class DeviceController {
         String manufacturer = request.getParameter("manufacturer");
         String model = request.getParameter("model");
         String ipAddress = request.getParameter("ipAddress");
-        int port = Integer.parseInt(request.getParameter("port"));
-        int xPos = Integer.parseInt(request.getParameter("xPos"));
-        int yPos = Integer.parseInt(request.getParameter("yPos"));
-        int rotation = Integer.parseInt(request.getParameter("rotation"));
+        int port;
+        try {
+            port = Integer.parseInt(request.getParameter("port"));
+            InetAddress.getByName(ipAddress);
+        } catch (UnknownHostException | NumberFormatException ex) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        int xPos;
+        try {
+            xPos = Integer.parseInt(request.getParameter("xPos"));
+        } catch (NumberFormatException ex) {
+            xPos = -1;
+        }
+        int yPos;
+        try {
+            yPos = Integer.parseInt(request.getParameter("yPos"));
+        } catch (NumberFormatException ex) {
+            yPos = -1;
+        }
+        int rotation;
+        try {
+            rotation = Integer.parseInt(request.getParameter("rotation"));
+        } catch (NumberFormatException ex) {
+            rotation = -1;
+        }
         DeviceType type = deviceTypeRepository.findByType(request.getParameter("type"));
         DeviceInfo info = deviceInfoRepository.findByManufacturerAndModelAndDeviceType(manufacturer, model, type);
-        Device device = new Device(info, ipAddress,port,xPos,yPos,rotation);
-        return new ResponseEntity<>(deviceRepository.save(device), HttpStatus.OK);
+        if(type != null && info != null) {
+            Device device = new Device(info, ipAddress, port, xPos, yPos, rotation);
+            return new ResponseEntity<>(deviceRepository.save(device), HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @RequestMapping(value = "/update"
-            , method = RequestMethod.POST
+            , method = RequestMethod.PUT
             , consumes = MediaType.APPLICATION_FORM_URLENCODED_VALUE)
     public ResponseEntity<Device> update(ServletRequest request) {
         int id = Integer.parseInt(request.getParameter("id"));
-        if(deviceRepository.findById(id).isPresent()) {
+        if (deviceRepository.findById(id).isPresent()) {
             Device device = deviceRepository.findById(id).get();
             String manufacturer = request.getParameter("manufacturer");
             String model = request.getParameter("model");
@@ -105,15 +138,25 @@ public class DeviceController {
         return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
     }
 
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @RequestMapping(value = "/remove"
-            , method = RequestMethod.POST
+            , method = RequestMethod.DELETE
             , consumes = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<String> removeOne(@RequestBody Device device) {
-        deviceRepository.delete(device);
-        return new ResponseEntity<>("Removed device: "
-                + device.getId(), HttpStatus.OK);
+    public ResponseEntity<String> removeOne(@RequestParam ("id") int id) {
+        try {
+            if (deviceRepository.findById(id).isPresent()) {
+                Device device = deviceRepository.findById(id).get();
+
+                deviceRepository.delete(device);
+                return new ResponseEntity<>("Removed device: "
+                        + device.getId(), HttpStatus.OK);
+            }
+        } catch (NumberFormatException ex) {
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
+    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @RequestMapping(value = "/removelist"
             , method = RequestMethod.POST
             , consumes = MediaType.APPLICATION_JSON_VALUE)
